refactor: split task service responsibilities
